# core-common-product-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-product-mgmt
  description: Product Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/products
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/products/{productId}:
    get:
      tags:
        - Product
      summary: Get product by ID
      description: Retrieve a specific product by its unique identifier
      operationId: getProductById
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved the product
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Product not found
    put:
      tags:
        - Product
      summary: Update product
      description: Update the information of an existing product by its unique identifier
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        required: true
      responses:
        '200':
          description: Product successfully updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          description: Invalid product data
        '404':
          description: Product not found
    delete:
      tags:
        - Product
      summary: Delete product
      description: Remove an existing product by its unique identifier
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Product successfully deleted
        '404':
          description: Product not found
...
  /api/v1/products/{productId}/documentation/{docId}:
    get:
      tags:
        - Product Documentation
      summary: Get product documentation by ID
      description: Retrieve a specific documentation item by its unique identifier
      operationId: getDocumentationById
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: docId
          in: path
          description: Unique identifier of the documentation item
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved the product documentation
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDocumentationDTO'
        '404':
          description: Product documentation not found
    put:
      tags:
        - Product Documentation
      summary: Update product documentation
      description: Update an existing documentation item for a specific product
      operationId: updateDocumentation
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: docId
          in: path
          description: Unique identifier of the documentation item
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDocumentationDTO'
        required: true
      responses:
        '200':
          description: Product documentation successfully updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDocumentationDTO'
        '400':
          description: Invalid documentation data
        '404':
          description: Product documentation not found
    delete:
      tags:
        - Product Documentation
      summary: Delete product documentation
      description: Remove an existing product documentation item by its unique identifier
      operationId: deleteDocumentation
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: docId
          in: path
          description: Unique identifier of the documentation item
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Product documentation successfully deleted
        '404':
          description: Product documentation not found
...
  /api/v1/products/{productId}/documentation:
    post:
      tags:
        - Product Documentation
      summary: Create product documentation
      description: Create a new documentation item for a specific product
      operationId: createDocumentation
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDocumentationDTO'
        required: true
      responses:
        '201':
          description: Product documentation successfully created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDocumentationDTO'
        '400':
          description: Invalid documentation data
        '404':
          description: Product not found
  /api/v1/products/{productId}/documentation/filter:
    post:
      tags:
        - Product Documentation
      summary: Filter product documentation
      description: Retrieve a paginated list of documentation items for a product based on filtering criteria
      operationId: filterDocumentation
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestProductDocumentationDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved filtered documentation
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
        '404':
          description: Product not found
...
  /api/v1/products/filter:
    post:
      tags:
        - Product
      summary: Filter products
      description: Retrieve a paginated list of products based on filtering criteria
      operationId: filterProducts
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestProductDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved filtered products
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
...
components:
  schemas:
    ProductDTO:
      required:
        - tenantId
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        productId:
          type: string
          format: uuid
          readOnly: true
        tenantId:
          type: string
          description: Unique identifier of the tenant instance this product belongs to
          format: uuid
        productCategoryId:
          type: string
          format: uuid
        productType:
          type: string
          enum:
            - FINANCIAL
            - NON_FINANCIAL
        productName:
          type: string
        productCode:
          type: string
        productDescription:
          type: string
        productStatus:
          type: string
          enum:
            - PROPOSED
            - ACTIVE
            - RETIRED
            - DRAFT
        launchDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
...
    ProductDocumentationDTO:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        productDocumentationId:
          type: string
          format: uuid
          readOnly: true
        productId:
          type: string
          format: uuid
        docType:
          type: string
          enum:
            - TNC
            - BROCHURE
            - POLICY_DOC
        documentManagerRef:
          type: integer
          format: int64
        dateAdded:
          type: string
          format: date-time
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestProductDocumentationDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/ProductDocumentationDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for documentation
...
    FilterRequestProductDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/ProductDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for products
...